==> Predictions/8_build_genomics_pdl1_df.ipynb <==
import pandas as pd
import helper_data_load as helper

# PDL1- Creat df for R
## Notes:
- Binary class for pdl1: 
    - patients with "1-49%", "<1%" -> low -> 0
    - patients with ">50%" -> high -> 1
### Results on datasets:
- CHUM
    - total: 61 patients
    - low -> 0 -> 39
    - high -> 1 -> 32
- IUCPQ
    - total: 54 patients
    - low -> 0 -> 15
    - high -> 1 -> 39

def get_pdl1_class_data(clinical, features_df):
    path='../../Data/'

    ch_cl=pd.read_csv(path + clinical + '.csv')
    chum_gen=pd.read_csv(path + features_df + '.csv')
    chum_gen = helper.prep_genomics(chum_gen)
    genetic_features = chum_gen.columns[1:]
    new_d=pd.merge(ch_cl,chum_gen, on='oncotech_id')
    print(len(new_d))
    new_d.dropna(subset=["pdl1_group"], inplace=True)
    new_d.reset_index(inplace=True)

    new_d["class_pdl1"] = 4
    new_d["class_pdl1"][new_d.pdl1_group.isin(["1-49%", "<1%"])] = 0
    new_d["class_pdl1"][new_d.pdl1_group.isin([">50%"])] = 1
    print(new_d["class_pdl1"].value_counts())
    return new_d, genetic_features


chum_gen_pdl1, genetic_features = get_pdl1_class_data("clinical_CHUM", "Discovery-GenomicsTop5000-March22023")
chum_gen_pdl1[list(genetic_features)+["class_pdl1"]].to_csv("../../Data/CHUM_genetics_pdl1_class.csv")

iucpq_gen_pdl1, genetic_features = get_pdl1_class_data("clinical_IUCPQ", "ValidationIUCPQ-March2 2023")
iucpq_gen_pdl1[list(genetic_features)+["class_pdl1"]].to_csv("../../Data/IUCPQ_genetics_pdl1_class.csv")

# PFS - Creat df for R

def get_pfs_class_data(clinical, features_df):
    path='../../Data/'

    ch_cl=pd.read_csv(path + clinical + '.csv')
    chum_gen=pd.read_csv(path + features_df + '.csv')
    chum_gen = helper.prep_genomics(chum_gen)
    genetic_features = chum_gen.columns[1:]
    new_d=pd.merge(ch_cl,chum_gen, on='oncotech_id')
    print(len(new_d))
    new_d.dropna(subset=["pfs_days"], inplace=True)
    new_d.reset_index(inplace=True)

    new_d["class_pfs"] = 4
    new_d["class_pfs"][new_d.pfs_days <= new_d.pfs_days.median()] = 0
    new_d["class_pfs"][new_d.pfs_days > new_d.pfs_days.median()] = 1
    print(new_d["class_pfs"].value_counts())
    return new_d, genetic_features


chum_gen_pfs, genetic_features = get_pfs_class_data("clinical_CHUM", "Discovery-GenomicsTop5000-March22023")
chum_gen_pfs[list(genetic_features)+["class_pfs"]].to_csv("../../Data/CHUM_genetics_pfs_class.csv")

iucpq_gen_pfs, genetic_features = get_pfs_class_data("clinical_IUCPQ", "ValidationIUCPQ-March2 2023")
iucpq_gen_pfs[list(genetic_features)+["class_pfs"]].to_csv("../../Data/IUCPQ_genetics_pfs_class.csv")

# OS - Creat df for R

def get_os_class_data(clinical, features_df):
    path='../../Data/'

    ch_cl=pd.read_csv(path + clinical + '.csv')
    chum_gen=pd.read_csv(path + features_df + '.csv')
    chum_gen = helper.prep_genomics(chum_gen)
    genetic_features = chum_gen.columns[1:]
    new_d=pd.merge(ch_cl,chum_gen, on='oncotech_id')
    print(len(new_d))
    new_d.dropna(subset=["os_days"], inplace=True)
    new_d.reset_index(inplace=True)

    new_d["class_os"] = 4
    new_d["class_os"][new_d.os_days <= new_d.os_days.median()] = 0
    new_d["class_os"][new_d.os_days > new_d.os_days.median()] = 1
    print(new_d["class_os"].value_counts())
    return new_d, genetic_features


chum_gen_os, genetic_features = get_os_class_data("clinical_CHUM", "Discovery-GenomicsTop5000-March22023")
chum_gen_os[list(genetic_features)+["class_os"]].to_csv("../../Data/CHUM_genetics_os_class.csv")

iucpq_gen_os, genetic_features = get_os_class_data("clinical_IUCPQ", "ValidationIUCPQ-March2 2023")
iucpq_gen_os[list(genetic_features)+["class_os"]].to_csv("../../Data/IUCPQ_genetics_os_class.csv")

# Distributions of enpoins in datasets

endpoint = "class_pdl1"
print("iucpq")
print(iucpq_gen_pdl1[endpoint].value_counts())
print("CHUM")
print(chum_gen_pdl1[endpoint].value_counts())

endpoint = "class_pfs"
print("iucpq")
print(iucpq_gen_pfs[endpoint].value_counts())
print("CHUM")
print(chum_gen_pfs[endpoint].value_counts())

endpoint = "class_os"
print("iucpq")
print(iucpq_gen_os[endpoint].value_counts())
print("CHUM")
print(chum_gen_os[endpoint].value_counts())
